+++ react/__mocks__/vtex.render-runtime.js
@@ -27,6 +27,7 @@ export const useRuntime = () => {
           },
         }
       }
+
       return {}
     },
   }